1、修改了一些数据库字段的类型 2、开始完成checkCampgroundList接口

+++ src/main/java/com/example/rv/pojo/VehicleMaintenance.java
@@ -9,7 +9,7 @@ public class VehicleMaintenance {
   private int vehicleMaintenanceVehicleId;
   private String vehicleMaintenanceMaintenanceDetails;
   private java.sql.Timestamp vehicleMaintenanceCompletedAt;
-  private int vehicleMaintenanceStatus;
+  private short vehicleMaintenanceStatus;
   private java.sql.Timestamp vehicleMaintenanceCreatedAt;
   private java.sql.Timestamp vehicleMaintenanceUpdatedAt;
